feat: Create a GlobalExceptionHandler and refactor DeleteCourse
Create the GlobalExceptionHandler to stardardize the CourseNotFoundException as a 404 return

+++ src/main/java/dev/marcostavares/dev_school/modules/course/services/DeleteCourse.java
@@ -5,6 +5,7 @@ import java.util.UUID;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import dev.marcostavares.dev_school.exceptions.CourseNotFoundException;
 import dev.marcostavares.dev_school.modules.course.repository.CourseRepository;
 
 @Service
@@ -14,6 +15,10 @@ public class DeleteCourse {
     private CourseRepository courseRepository;
 
     public String execute(UUID id) {
+        if (!courseRepository.existsById(id)) {
+            throw new CourseNotFoundException();
+        }
+
         courseRepository.deleteById(id);
         return "Course successfully deleted";
     }
